greybus: audio: Add Greybus Audio Device Class Protocol Driver
Add a driver that supports the Greybus Audio Device Class Protocol.
The driver handles incoming Greybus audio messages sent by the AP
and sends audio events to the AP when the occur.  The driver configures
an I2S connection on the GPBridge so it can send and receive audio data
to/from an audio codec.  The audio codec is also controlled by this driver
(via the appropriate codec driver) to manage the other end of the I2S
connection.

The Audio Module topology data defined by the protocol is provided by
the codec driver.  This is a layering violation but it is the only
sensible way to support an unknown set of audio codecs.

A Greybus Audio Module has at least one Audio Bundle.  Each Audio Bundle
represents an audio codec and has exactly one Audio Management CPort and
one or more Audio Data CPorts.  Each Audio Data CPort corresponds to a
usable I2S connection on the associated codec.  An I2S connection is also
known as a Digital Audio Interface (DAI).

Audio Bundle information is specified in the module's manifest but there
is no easy way for drivers on the module to access the manifest data.
Instead, the audio_board driver is used which provides a convenient way
to retrieve the same information.

Testing Done: Played audio using the White Audio Module
Bug Number: SW-1060
Review URL:
Reviewed by:

Signed-off-by: Chris Myerchin <myerchin_chris@projectara.com>
Signed-off-by: Mark Greer <mgreer@animalcreek.com>